requirements.txt
# Core Google Vertex AI SDK
google-cloud-aiplatform>=1.44.0
vertexai>=0.1.0

# Generative AI (Gemini)
google-generativeai>=0.3.2

# GCP authentication and API support
google-auth==2.29.0
google-api-core>=2.17.0
protobuf!=4.21.0
grpcio==1.62.1

# BigQuery support
google-cloud-bigquery>=3.14.0

# UI - Streamlit
streamlit>=1.35.0

# Environment variable support
python-dotenv==1.0.1

# Data manipulation
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3.0

# Plotting and visualization
plotly==5.22.0

# Text processing
textblob

# Embedding / Vector Search (optional local)
scipy
faiss-cpu

# YAML support
PyYAML>=6.0.1,<6.1

# Logging and progress
loguru
tqdm

# Testing tools
pytest
pytest-mock

# Dev tools (optional)
black
flake8
